requirements.txt
Pillow==9.5.0
pytube==15.0.0
ttkbootstrap==1.10.1